Warning compilation
Former-commit-id: 73f6dca613de498d57e622231260c5f40770a033

+++ libraries/reconstruction/ml_align2d.h
@@ -92,7 +92,7 @@ public:
     double wsum_corr, sum_refw, sum_refw2, maxweight;
     double wsum_sc, wsum_sc2, wsum_offset, old_bgmean;
     double mindiff;
-    int sigdim;
+    size_t sigdim;
     int ioptx, iopty;
     std::vector<int> ioptx_ref, iopty_ref, ioptflip_ref;
     /** Taken from PreselectFastSignificant. */
